--- admin/src/Layout.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { Outlet } from "react-router-dom";
-import Navbar from "./navbar";
-
-const Layout = () => {
-  return (
-    <>
-      <Outlet />
-    </>
-  );
-};
-
-export default Layout;
