--- admin/src/navbar.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { BrowserRouter, Route, Link } from "react-router-dom";
-
-function Navbar() {
-  return (
-    <nav>
-      <ul>
-        <li>
-          <Link to="/">Sıgn in</Link>
-        </li>
-        <li>
-          <Link to="/signup">Sıgn up</Link>
-        </li>
-      </ul>
-    </nav>
-  );
-}
-
-export default Navbar;
