+++ client/src/Auth/Admin/AdminLogin.js
@@ -1,46 +1,76 @@
 import * as React from "react";
-import Avatar from "@mui/material/Avatar";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import Checkbox from "@mui/material/Checkbox";
+import {useNavigate} from "react-router-dom";
 import Paper from "@mui/material/Paper";
-import { BrowserRouter, Route, Link } from "react-router-dom";
 import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
+import {createTheme, ThemeProvider} from "@mui/material/styles";
 import axios from "axios";
 
 const theme = createTheme();
 
-export default function CreateTeamAccount() {
+export default function AdminLogin() {
+  const navigate = useNavigate();
   const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
     console.log("button pressed");
+    if (!data.get("password")) {
+      alert("Please enter your password.");
+      return;
+    }
     axios
-      /*.post("/api/v1/users", {
-        email: data.get("email"),
-        password: data.get("password"),
-      })*/
-      .post("/api/v1", {
-        email: data.get("email"),
-        password: data.get("password"),
-        _collection: "team_accounts"
+      .get("/api/v1", {
+          params: {
+              _collection: "admin",
+              email: "admin@tff.com",
+          },
       })
       .catch((err) => {
-        console.log(err);
+          console.log(err);
       })
       .then((res) => {
-        console.log(res);
+          if (
+              res &&
+              (res.data.items.length === 0 ||
+                  res.data.items[0].password !== data.get("password"))
+          ) {
+              alert("Invalid password.");
+              return;
+          }
+          if (res) {
+              navigate("/adminpanel", { state: res.data.items[0] });
+          } else {
+              alert("There was an error. Please try again.");
+          }
       });
   };
 
   return (
     <ThemeProvider theme={theme}>
+      <Grid container component="main" sx={{ height: "100vh" }}>
         <CssBaseline />
+        <Grid
+          item
+          xs={false}
+          sm={4}
+          md={7}
+          sx={{
+            backgroundImage:
+              "url(https://www.evrensel.net/upload/dosya/167821.jpg)",
+            backgroundRepeat: "no-repeat",
+            backgroundColor: (t) =>
+              t.palette.mode === "light"
+                ? t.palette.grey[50]
+                : t.palette.grey[900],
+            backgroundSize: "cover",
+            backgroundPosition: "center",
+          }}
+        />
+        <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
           <Box
             sx={{
               my: 8,
@@ -51,7 +81,7 @@ export default function CreateTeamAccount() {
             }}
           >
             <Typography component="h1" variant="h5">
-              Create Team Account
+              RASA Admin Panel
             </Typography>
             <Box
               component="form"
@@ -59,16 +89,6 @@ export default function CreateTeamAccount() {
               onSubmit={handleSubmit}
               sx={{ mt: 1 }}
             >
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                autoComplete="email"
-                autoFocus
-              />
               <TextField
                 margin="normal"
                 required
@@ -85,10 +105,12 @@ export default function CreateTeamAccount() {
                 variant="contained"
                 sx={{ mt: 3, mb: 2 }}
               >
-                Set Up Team Account
+                Enter Admin Panel
               </Button>
             </Box>
           </Box>
+        </Grid>
+      </Grid>
     </ThemeProvider>
   );
 }
