+++ client/src/Auth/Team/TeamLogin.js
@@ -1,50 +1,51 @@
 import * as React from "react";
-import Avatar from "@mui/material/Avatar";
 import Button from "@mui/material/Button";
 import CssBaseline from "@mui/material/CssBaseline";
 import TextField from "@mui/material/TextField";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import Checkbox from "@mui/material/Checkbox";
-//import Link from "@mui/material/Link";
-import { BrowserRouter, Route, Link } from "react-router-dom";
-import { useNavigate } from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import Paper from "@mui/material/Paper";
 import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
-import { createTheme, ThemeProvider } from "@mui/material/styles";
+import {createTheme, ThemeProvider} from "@mui/material/styles";
 import axios from "axios";
 
 const theme = createTheme();
 
-export default function AdminLogin() {
+export default function TeamLogin() {
   const navigate = useNavigate();
   const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
     console.log("button pressed");
-    if (/*!data.get("email") || */!data.get("password")) {
+    if (!data.get("email") || !data.get("password")) {
+        alert("Invalid email or password");
       return;
     }
     axios
-      .get("/api/v1/users", {
-        params: {
-          email: "rasaAdmin@tff.com"/*data.get("email")*/,
-        },
+      .get("/api/v1", {
+          params: {
+              _collection: "team_accounts",
+              email: data.get("email"),
+          },
       })
       .catch((err) => {
-        console.log(err);
+          console.log(err);
       })
       .then((res) => {
-        console.log(res);
-        if (
-          res.data.users.length == 0 ||
-          res.data.users[0].password != data.get("password")
-        ) {
-          return;
-        }
-        console.log(res.data.users[0]);
-        navigate("/adminpanel", { state: res.data.users[0] });
+          if (
+              res &&
+              (res.data.items.length === 0 ||
+                  res.data.items[0].password !== data.get("password"))
+          ) {
+              alert("Invalid email or password.");
+              return;
+          }
+          if (res) {
+                navigate("/teampanel", { state: res.data.items[0] });
+          } else {
+            alert("There was an error. Please try again.");
+          }
       });
   };
 
@@ -80,7 +81,7 @@ export default function AdminLogin() {
             }}
           >
             <Typography component="h1" variant="h5">
-              RASA Admin Panel
+              RASA Team Panel Sign In
             </Typography>
             <Box
               component="form"
@@ -88,6 +89,16 @@ export default function AdminLogin() {
               onSubmit={handleSubmit}
               sx={{ mt: 1 }}
             >
+              <TextField
+                margin="normal"
+                required
+                fullWidth
+                id="email"
+                label="Team Email Address"
+                name="email"
+                autoComplete="email"
+                autoFocus
+              />
               <TextField
                 margin="normal"
                 required
@@ -104,7 +115,7 @@ export default function AdminLogin() {
                 variant="contained"
                 sx={{ mt: 3, mb: 2 }}
               >
-                Enter Admin Panel
+                Enter Team Panel
               </Button>
             </Box>
           </Box>
