+++ package.json
@@ -5,25 +5,19 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build-client": "cd client && npm run build",
-    "build-team": "cd team && npm run build",
-    "build-admin": "cd admin && npm run build",
+    "build": "cd client && npm run build",
     "install-client": "cd client && npm install",
-    "install-team": "cd team && npm install",
-    "install-admin": "cd admin && npm install",
-    "heroku-postbuild": "npm run install-client && npm run install-team && npm run install-admin && npm run build-client && npm run build-team && npm run build-team",
+    "heroku-postbuild": "npm run install-client && npm run build",
     "start": "node server.js",
     "client": "cd client && npm start",
-    "admin": "cd admin && npm start",
-    "team": "cd team && npm start",
-    "dev": "concurrently -n server,admin,client,team -c 'red,green,blue,yellow' \"nodemon server.js\" \"npm run admin\" \"npm run client\" \"npm run team\""
+    "dev": "concurrently -n server,client -c 'red,green'  \"nodemon server.js\" \"npm run client\""
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "axios": "^1.1.3",
-    "concurrently": "^7.6.0",
+    "concurrently": "^7.5.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
